requirements.txt
flask
python-dotenv
replicate